Adding "mutable module data" to `example.f90`.
A "knob" that can be turned, if you will.

Also updated `journal-linux-3.6.txt` because I dropped the
`conda` installed `gfortran` on my Linux machine so `pyenv` was
no longer jumping in front of the line.

Also updated to Cython 0.26.1 for the generated `fast.c`.

ASIDE: `make run-golang` seems to be broken. May be caused by
me upgrading from Go 1.8 to 1.9 on my machine?