Finished porting over the `cython` example.
At this point all of the original README files are gone. There is still
plenty to do (e.g. the Cython example should explain how things are
done in `setup.py`). But this puts the project in a new place where
"everything" is happening in documentation rather than spread in
other files.